Move frontend nav components to new directory
