+++ resources/js/components/frontend/nav/nav-bar.tsx
